fix errors in the HR Addon Setting

+++ hr_addon/hr_addon/doctype/hr_addon_settings/hr_addon_settings.py
@@ -10,9 +10,10 @@ class HRAddonSettings(Document):
 	def before_save(self):
 		# remove the old ics file
 		old_doc = self.get_doc_before_save()
-		old_file_name = old_doc.name_of_calendar_export_ics_file
-		if old_file_name != self.name_of_calendar_export_ics_file:
-			os.remove("{}/public/files/{}.ics".format(frappe.utils.get_site_path(), old_file_name))
+		if old_doc:
+			old_file_name = old_doc.name_of_calendar_export_ics_file
+			if old_file_name != self.name_of_calendar_export_ics_file:
+				os.remove("{}/public/files/{}.ics".format(frappe.utils.get_site_path(), old_file_name))
 
 		# remove also the Urlaubskalender.ics, if exist
 		if os.path.exists("{}/public/files/Urlaubskalender.ics".format(frappe.utils.get_site_path())):
